[Refactoring] changed double types to decimal for currency

+++ src/lunes.Application/UseCases/Accounts/AddExpense/AddExpenseOutput.cs
@@ -6,9 +6,9 @@ namespace lunes.Application.UseCases.Accounts.AddExpense
     {
 	    public Guid AccountId { get; set; }
 
-	    public double Balance { get; }
+	    public decimal Balance { get; }
 
-	    public AddExpenseOutput(Guid accountId, double balance)
+	    public AddExpenseOutput(Guid accountId, decimal balance)
 	    {
 		    AccountId = accountId;
 		    Balance = balance;
