+++ src/lunes.Application/UseCases/Accounts/AddRevenue/AddRevenueOutput.cs
@@ -6,9 +6,9 @@ namespace lunes.Application.UseCases.Accounts.AddRevenue
     {
 		public Guid AccountId { get; set; }
 
-		public double Balance { get; }
+		public decimal Balance { get; }
 
-		public AddRevenueOutput(Guid accountId, double balance)
+		public AddRevenueOutput(Guid accountId, decimal balance)
 		{
 			AccountId = accountId;
 			Balance = balance;
