+++ src/lunes.Application/UseCases/Accounts/MakeTransfer/IMakeTransferUseCase.cs
@@ -5,7 +5,7 @@ namespace lunes.Application.UseCases.Accounts.MakeTransfer
 {
     public interface IMakeTransferUseCase
     {
-	    Task<MakeTransferOutput> Run(string name, double amount, Guid fromAccountId, Guid toAccountId);
+	    Task<MakeTransferOutput> Run(string name, decimal amount, Guid fromAccountId, Guid toAccountId);
 
     }
 }
